no needs to have libft

+++ src/utils/split.c
@@ -3,20 +3,20 @@
 /*                                                        :::      ::::::::   */
 /*   split.c                                            :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: lgaume <lgaume@student.42.fr>              +#+  +:+       +#+        */
+/*   By: lgaume <lgaume@student.42lausanne.ch>      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/11/17 05:59:51 by lgaume            #+#    #+#             */
-/*   Updated: 2023/11/17 05:59:59 by lgaume           ###   ########.fr       */
+/*   Updated: 2023/11/19 08:41:13 by lgaume           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../../inc/push_swap.h"
 
 static int	count_words(char *s, char c)
-{	
+{
 	int		count;
 	bool	inside_word;
-	
+
 	count = 0;
 	while (*s)
 	{
@@ -42,14 +42,14 @@ static char	*get_next_word(char *s, char c)
 	char		*next_word;
 	int			len;
 	int			i;
-	
+
 	len = 0;
 	i = 0;
 	while (s[cursor] == c)
 		++cursor;
 	while ((s[cursor + len] != c) && s[cursor + len])
 		++len;
-	next_word = malloc((size_t)len * sizeof(char) + 1);
+	next_word = malloc(len * sizeof(char) + 1);
 	if (!next_word)
 		return (NULL);
 	while ((s[cursor] != c) && s[cursor])
@@ -63,12 +63,12 @@ char	**split(char *s, char c)
 	int		words_count;
 	char	**output;
 	int		i;
-	
+
 	i = 0;
 	words_count = count_words(s, c);
 	if (!words_count)
 		exit(1);
-	output = malloc(sizeof(char *) * (size_t)(words_count + 2));
+	output = malloc(sizeof(char *) * (words_count + 2));
 	if (!output)
 		return (NULL);
 	while (words_count-- >= 0)
